FileManagement implemented.  Bug fixes. Code updates.
appendDataToFile(), showFileData(), createFile() implemented and tested.
Several parts of the code fixed: some null checkers added.
The main change is in getChild() which made an automatic cast to
Directory, now it return a FileSystemEntitiy. Almost no code required
change, since all of the functions in FileSystemManager made checks to
make sure that the returned type is the expected one (i.e. that a
Directory is returned from getChild()). The only problematic situation is
in the poof.core.Parses, which will require some additional type checking.